initial working commit
Add a script to generate random numbers for different lotteries. It can
be used for PowerBall, Mega Millions, and Carolina Cash 5. The script
takes the name of a lottery and a number as arguments. The number is
the total set of numbers to generate, as in the number of tickets you
want to play.